+++ Sources/PackageCollections/JSONModel/JSONCollection+v1.swift
@@ -250,116 +250,3 @@ extension JSONPackageCollectionModel.V1 {
         }
     }
 }
-
-// MARK: - Validations
-
-extension JSONPackageCollectionModel.V1 {
-    public struct Validator {
-        public let configuration: Configuration
-
-        public init(configuration: Configuration = .init()) {
-            self.configuration = configuration
-        }
-
-        public func validate(collection: Collection) -> [ValidationMessage]? {
-            var messages = [ValidationMessage]()
-
-            let packages = collection.packages
-            // Stop validating if collection doesn't pass basic checks
-            if packages.isEmpty {
-                messages.append(.error("A collection must contain at least one package.", property: "packages"))
-            } else if packages.count > self.configuration.maximumPackageCount {
-                messages.append(.warning("The collection has (\(packages.count)) packages, which is more than the recommended maximum (\(self.configuration.maximumPackageCount)) and extra data might be ignored.", property: "packages"))
-            } else {
-                packages.forEach { self.validate(package: $0, messages: &messages) }
-            }
-
-            guard messages.isEmpty else {
-                return messages
-            }
-
-            return nil
-        }
-
-        // TODO: validate package url?
-        private func validate(package: Collection.Package, messages: inout [ValidationMessage]) {
-            let packageID = PackageIdentity(url: package.url.absoluteString).description
-
-            // Check for duplicate versions
-            let nonUniqueVersions = Dictionary(grouping: package.versions, by: { $0.version }).filter { $1.count > 1 }.keys
-            if !nonUniqueVersions.isEmpty {
-                messages.append(.error("Duplicate version(s) found in package \(packageID): \(nonUniqueVersions).", property: "package.versions"))
-            }
-
-            var nonSemanticVersions = [String]()
-            let semanticVersions: [TSCUtility.Version] = package.versions.compactMap {
-                let semver = TSCUtility.Version(string: $0.version)
-                if semver == nil {
-                    nonSemanticVersions.append($0.version)
-                }
-                return semver
-            }
-
-            guard nonSemanticVersions.isEmpty else {
-                messages.append(.error("Non semantic version(s) found in package \(packageID): \(nonSemanticVersions).", property: "package.versions"))
-                // The next part of validation requires sorting the semvers. Cannot continue if non-semver.
-                return
-            }
-
-            let sortedVersions = semanticVersions.sorted(by: >)
-
-            var currentMajor: Int?
-            var majorCount = 0
-            var minorCount = 0
-            for version in sortedVersions {
-                if version.major != currentMajor {
-                    currentMajor = version.major
-                    majorCount += 1
-                    minorCount = 0
-                }
-
-                guard majorCount <= self.configuration.maximumMajorVersionCount else {
-                    messages.append(.warning("Package \(packageID) includes too many major versions. Only \(self.configuration.maximumMajorVersionCount) is allowed and extra data might be ignored.", property: "package.versions"))
-                    break
-                }
-                guard minorCount < self.configuration.maximumMinorVersionCount else {
-                    // !-safe currentMajor cannot be nil at this point
-                    messages.append(.warning("Package \(packageID) includes too many minor versions for major version \(currentMajor!). Only \(self.configuration.maximumMinorVersionCount) is allowed and extra data might be ignored.", property: "package.versions"))
-                    break
-                }
-
-                minorCount += 1
-            }
-
-            package.versions.forEach { version in
-                if version.products.isEmpty {
-                    messages.append(.error("Package \(packageID) version \(version.version) does not contain any products.", property: "version.products"))
-                }
-                version.products.forEach { product in
-                    if product.targets.isEmpty {
-                        messages.append(.error("Product \(product.name) of package \(packageID) version \(version.version) does not contain any targets.", property: "product.targets"))
-                    }
-                }
-
-                if version.targets.isEmpty {
-                    messages.append(.error("Package \(packageID) version \(version.version) does not contain any targets.", property: "version.targets"))
-                }
-            }
-        }
-
-        public struct Configuration {
-            public var maximumPackageCount: Int
-            public var maximumMajorVersionCount: Int
-            public var maximumMinorVersionCount: Int
-
-            public init(maximumPackageCount: Int? = nil,
-                        maximumMajorVersionCount: Int? = nil,
-                        maximumMinorVersionCount: Int? = nil) {
-                // TODO: where should we read defaults from?
-                self.maximumPackageCount = maximumPackageCount ?? 50
-                self.maximumMajorVersionCount = maximumMajorVersionCount ?? 2
-                self.maximumMinorVersionCount = maximumMinorVersionCount ?? 3
-            }
-        }
-    }
-}
